backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.32.3
python-dotenv==1.0.0
pydantic==2.8.2
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langgraph==0.3.27